refactor(payment): PAYPAL-3603 updated braintree analytics folder name and codeowners file
