+++ .github/CODEOWNERS
@@ -64,8 +64,8 @@
 
 ## PayPal team
 
-/packages/core/src/analytics/braintree-connect-tracker
-/packages/core/src/analytics/paypal-commerce-connect-tracker
+/packages/core/src/analytics/braintree-analytic-tracker @bigcommerce/team-paypal
+/packages/core/src/analytics/paypal-commerce-analytic-tracker @bigcommerce/team-paypal
 /packages/core/src/checkout-buttons/strategies/braintree @bigcommerce/team-paypal
 /packages/core/src/checkout-buttons/strategies/paypal @bigcommerce/team-paypal
 /packages/core/src/customer/strategies/braintree @bigcommerce/team-paypal
