+++ packages/core/src/analytics/braintree-analytic-tracker/index.ts
@@ -1,0 +1,2 @@
+export { default as createBraintreeAnalyticTracker } from './create-braintree-analytic-tracker';
+export { default as BraintreeAnalyticTrackerService } from './braintree-analytic-tracker-service';
