--- packages/core/src/analytics/braintree-connect-tracker/index.ts
@@ -1,3 +1,0 @@
-// TODO: variables names should be updated with actual naming (Analytic instead of Connect)
-export { default as createBraintreeConnectTracker } from './create-braintree-analytic-tracker';
-export { default as BraintreeConnectTrackerService } from './braintree-analytic-tracker-service';
