+++ packages/core/src/analytics/index.ts
@@ -2,9 +2,9 @@ export { default as createStepTracker } from './create-step-tracker';
 export { default as StepTracker } from './step-tracker';
 
 export {
-    createBraintreeConnectTracker,
-    BraintreeConnectTrackerService,
-} from './braintree-connect-tracker';
+    createBraintreeAnalyticTracker,
+    BraintreeAnalyticTrackerService,
+} from './braintree-analytic-tracker';
 
 export {
     createPayPalCommerceAnalyticTracker,
